session: added filters and write/lookup error parsing

+++ elliptics/errors.go
@@ -27,7 +27,7 @@ import (
 
 type DnetError struct {
 	Code		int
-	Flags		int
+	Flags		uint64
 	Message		string
 }
